# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Api SDU-JobQuest",
    "version": "1.0.0",
    "description": ""
  },
  "host": "localhost:8800",
  "basePath": "/",
  "schemes": [
    "https",
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/save/user/": {
      "get": {
        "tags": [
          "Lưu công việc"
        ],
        "description": "",
        "parameters": [
          {
            "name": "idJob",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/apply/user": {
      "get": {
        "tags": [
          "Ứng tuyển"
        ],
        "description": "",
        "parameters": [
          {
            "name": "idJob",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}